14/6/24 Items fixed

+++ app/src/main/java/com/matiasd/brainlist3/Lists.kt
@@ -18,8 +18,12 @@ import com.google.firebase.auth.FirebaseAuth
 import android.content.res.ColorStateList
 import android.graphics.Color
 import android.view.MenuItem
+import android.widget.Toast
 import androidx.appcompat.app.ActionBarDrawerToggle
+import androidx.appcompat.app.AlertDialog
 import androidx.core.view.GravityCompat
+import androidx.core.view.marginBottom
+import androidx.core.view.setMargins
 import androidx.drawerlayout.widget.DrawerLayout
 import com.google.android.material.navigation.NavigationView
 import com.google.firebase.auth.ktx.auth
@@ -83,21 +87,29 @@ class Lists : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListen
         val linearLayout = scrollView.findViewById<LinearLayout>(R.id.lLItems) // Obtén el LinearLayout dentro del ScrollView
 
         for (item in dataList) {
+            val containerLayout = LinearLayout(this)
+            containerLayout.orientation = LinearLayout.VERTICAL
+
             val horizontalLayout = LinearLayout(this)
             horizontalLayout.orientation = LinearLayout.HORIZONTAL
+            horizontalLayout.setBackgroundResource(R.drawable.button_list)
 
-            val params = LinearLayout.LayoutParams(
+            val layoutparams = LinearLayout.LayoutParams(
+                LinearLayout.LayoutParams.MATCH_PARENT,
+                125
+            )
+            layoutparams.setMargins(0, 10, 0, 0)
+            horizontalLayout.layoutParams = layoutparams
+
+            val checkboxparams = LinearLayout.LayoutParams(
                 LinearLayout.LayoutParams.WRAP_CONTENT,
                 LinearLayout.LayoutParams.WRAP_CONTENT
             )
 
             val checkBox = CheckBox(this)
-            checkBox.gravity = Gravity.CENTER_VERTICAL
+            checkBox.layoutParams = checkboxparams // Establecer el ancho de CheckBox como WRAP_CONTENT
 
-            checkBox.layoutParams = params // Establecer el ancho de CheckBox como WRAP_CONTENT
-
-            //COLOR DEL CHECKBOX
-
+            // COLOR DEL CHECKBOX
             val colorStateList = ColorStateList(
                 arrayOf(intArrayOf(android.R.attr.state_checked), intArrayOf(-android.R.attr.state_checked)),
                 intArrayOf(Color.RED, Color.GREEN) // Cambia Color.RED y Color.GREEN por los colores deseados
@@ -116,15 +128,81 @@ class Lists : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListen
                 LinearLayout.LayoutParams.WRAP_CONTENT,
                 1.0f
             )
+            weightParams.setMargins(0, 8, 0, 0)
             textView.layoutParams = weightParams
 
+            // Crear y configurar el ImageButton
+            val imageButtonParams = LinearLayout.LayoutParams(
+                LinearLayout.LayoutParams.WRAP_CONTENT,
+                LinearLayout.LayoutParams.WRAP_CONTENT
+            )
+            val imageButton = ImageButton(this)
+            imageButton.layoutParams = imageButtonParams
+            imageButton.setImageResource(android.R.drawable.ic_menu_close_clear_cancel) // Reemplaza con tu icono deseado
+            imageButton.setBackgroundColor(Color.TRANSPARENT) // Hacer el fondo transparente
+            imageButton.setOnClickListener {
+                // Crear el AlertDialog
+                val builder = AlertDialog.Builder(this, R.style.CustomAlertDialog)
+                builder.setTitle("Eliminar Item")
+                builder.setMessage("¿Desea eliminar este item?")
+                builder.setPositiveButton("Confirmar") { dialog, which ->
+                    // Acción al confirmar
+                    listdb.deleteSingleItem(userId, nombreLista, item.first)
+                    linearLayout.removeView(containerLayout)
+                    Toast.makeText(this, "Item eliminado", Toast.LENGTH_SHORT).show()
+                }
+                builder.setNegativeButton("Cancelar") { dialog, which ->
+                    // Acción al cancelar
+                    dialog.dismiss()
+                }
 
-            // Agregar CheckBox y TextView al diseño horizontal
+                // Mostrar el AlertDialog
+                builder.show()
+            }
+
+            // Agregar CheckBox, TextView y ImageButton al diseño horizontal
             horizontalLayout.addView(checkBox)
             horizontalLayout.addView(textView)
+            horizontalLayout.addView(imageButton)
 
-            // Agregar el diseño horizontal al diseño lineal principal
-            linearLayout.addView(horizontalLayout)
+            // Crear y configurar la TextView para la descripción
+            val descriptionTextView = TextView(this)
+            descriptionTextView.text = item.second
+            descriptionTextView.textSize = 20f
+            descriptionTextView.setTextColor(Color.GRAY)
+            descriptionTextView.setBackgroundResource(R.drawable.description_box)
+            descriptionTextView.setPadding(10, 5, 10, 5)
+            descriptionTextView.visibility = View.GONE // Inicialmente oculta
+
+            // Limitar el ancho máximo del descriptionTextView y centrarlo
+            val descriptionParams = LinearLayout.LayoutParams(
+                750,
+                LinearLayout.LayoutParams.WRAP_CONTENT
+            )
+            descriptionParams.setMargins(20, 0, 20, 0) // Margen opcional para darle espacio al texto
+            descriptionParams.gravity = Gravity.CENTER_HORIZONTAL // Centrar horizontalmente
+            descriptionTextView.layoutParams = descriptionParams
+            descriptionTextView.maxWidth = 600 // Establece el ancho máximo deseado en píxeles
+
+            // Configurar el click listener para el TextView del item
+            textView.setOnClickListener {
+                if(item.second != ""){
+                    if (descriptionTextView.visibility == View.GONE) {
+                        descriptionTextView.visibility = View.VISIBLE
+                    } else {
+                        descriptionTextView.visibility = View.GONE
+                    }
+                }else{
+                    descriptionTextView.visibility = View.GONE
+                }
+            }
+
+            // Agregar el diseño horizontal y la descripción al diseño contenedor
+            containerLayout.addView(horizontalLayout)
+            containerLayout.addView(descriptionTextView)
+
+            // Agregar el diseño contenedor al diseño lineal principal
+            linearLayout.addView(containerLayout)
         }
 
         val btnAdd = findViewById<ImageButton>(R.id.btnAddItem)
@@ -141,15 +219,25 @@ class Lists : AppCompatActivity(), NavigationView.OnNavigationItemSelectedListen
         //DELETE BOTON
         val btnDeleteList = findViewById<ImageButton>(R.id.btnDeleteList)
         btnDeleteList.setOnClickListener {
-            // Aquí puedes eliminar la lista usando su nombre (nombreLista)
-            // Por ejemplo, puedes llamar a una función que elimine la lista de tu base de datos
-            // Aquí asumimos que tienes una función en tu base de datos (listdb) que elimina la lista
-            listbuttondb.deleteList(userId, nombreLista)
-            listdb.deleteList(userId, nombreLista)
-            // Después de eliminar la lista, puedes regresar a la actividad anterior o realizar cualquier otra acción
-            val intent = Intent(this, MainMenu::class.java)
-            startActivity(intent)
-            finish() // Termina esta actividad para evitar que el usuario vuelva a ella usando el botón "Atrás"
+            // Crear el AlertDialog
+            val builder = AlertDialog.Builder(this, R.style.CustomAlertDialog)
+            builder.setTitle("Eliminar Lista")
+            builder.setMessage("¿Desea eliminar esta lista?")
+            builder.setPositiveButton("Confirmar") { dialog, which ->
+                // Acción al confirmar
+                listbuttondb.deleteList(userId, nombreLista)
+                listdb.deleteList(userId, nombreLista)
+                // Después de eliminar la lista, puedes regresar a la actividad anterior o realizar cualquier otra acción
+                val intent = Intent(this, MainMenu::class.java)
+                startActivity(intent)
+                finish() // Termina esta actividad para evitar que el usuario vuelva a ella usando el botón "Atrás"
+            }
+            builder.setNegativeButton("Cancelar") { dialog, which ->
+                // Acción al cancelar
+                dialog.dismiss()
+            }
+            // Mostrar el AlertDialog
+            builder.show()
         }
         //DELETE BOTON
     }
